requirements.txt
discord.py
python-dotenv
openai
requests
beautifulsoup4
scrapingant_client
